fix: code review fixes - PlayerStatsService null guards, performance, decimal types
- Add ArgumentNullException.ThrowIfNull(players) to GetTeamTotals and GetTeamAverages
- Refactor both methods from 7 LINQ passes to single foreach loop each
- Extract private ComputeEstimatedValue helper; use const decimal multiplier to prevent int overflow
- Change TeamAverages.AvgEstimatedValue from double to decimal for type consistency

+++ Hattrick.Core/Models/TeamAverages.cs
@@ -16,7 +16,7 @@ public sealed class TeamAverages
     /// <summary>
     /// Arithmetic mean of estimated market values. Formula: each player's TSI * 25.
     /// </summary>
-    public double AvgEstimatedValue { get; init; }
+    public decimal AvgEstimatedValue { get; init; }
 
     /// <summary>Arithmetic mean of player ages.</summary>
     public double AvgAge { get; init; }
